@@ -10,26 +10,38 @@
   },
   "dependencies": {
     "core-js": "^3.14.0",
-    "material-design-icons-iconfont": "^5.0.1",
-    "vue": "^2.6.13",
-    "vuetify": "^2.6.10"
+    "material-design-icons-iconfont": "^6.7.0",
+    "vue": "^3.5.13",
+    "vuetify": "^3.7.5"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^4.1.2",
-    "@vue/cli-plugin-eslint": "^4.1.2",
-    "@vue/cli-plugin-unit-mocha": "^4.1.2",
-    "@vue/cli-service": "^4.1.2",
-    "@vue/test-utils": "1.0.0-beta.31",
-    "babel-eslint": "^10.1.0",
+    "@babel/eslint-parser": "^7.28.5",
+    "@mdi/font": "^7.4.47",
+    "@vue/cli-plugin-babel": "^5.0.8",
+    "@vue/cli-plugin-eslint": "^5.0.8",
+    "@vue/cli-plugin-unit-mocha": "^5.0.8",
+    "@vue/cli-service": "^5.0.8",
+    "@vue/test-utils": "^2.4.6",
     "chai": "^4.3.4",
-    "eslint": "^5.16.0",
-    "eslint-plugin-vue": "^5.0.0",
-    "null-loader": "^3.0.0",
-    "sass": "~1.32.6",
-    "sass-loader": "^8.0.2",
-    "vue-cli-plugin-vuetify": "^2.0.3",
-    "vue-template-compiler": "^2.6.11",
-    "vuetify-loader": "^1.4.3",
-    "webpack": "^4.46.0"
+    "eslint": "^8.57.0",
+    "eslint-plugin-vue": "^9.27.0",
+    "sass": "^1.70.0",
+    "vite-plugin-vuetify": "^2.0.4",
+    "webpack": "^5.90.0"
+  },
+  "eslintConfig": {
+    "root": true,
+    "env": {
+      "node": true
+    },
+    "extends": [
+      "plugin:vue/vue3-essential",
+      "eslint:recommended"
+    ],
+    "parserOptions": {
+      "parser": "@babel/eslint-parser",
+      "requireConfigFile": false
+    },
+    "rules": {}
   }
 }
