@@ -1,11 +1,8 @@
-//import 'roboto-fontface/css/roboto/roboto-fontface.css';
-import Vue from "vue";
+import { createApp } from "vue";
 import App from "./App.vue";
 import vuetify from './plugins/vuetify';
+import 'material-design-icons-iconfont/dist/material-design-icons.css';
 
-Vue.config.productionTip = false;
-
-new Vue({
-  vuetify,
-  render: h => h(App)
-}).$mount("#app");
+createApp(App)
+  .use(vuetify)
+  .mount("#app");
